Random Hill Climbing
Progress is made slowly and randomly
The program grabs a random queen and moves it to a random square. If the
new configuration is better then the old the configurastion is keept.
This is an extremly slow solution and is only being commited so that I
can work on it on other computers.